+++ src/Server/Persistence/Systems/Persistence/PersistenceInternalController.cs
@@ -14,7 +14,11 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
+using Microsoft.Extensions.Logging;
+using Sovereign.EngineCore.Components;
 using Sovereign.EngineCore.Events;
+using Sovereign.EngineCore.Events.Details;
+using EventId = Sovereign.EngineCore.Events.EventId;
 
 namespace Sovereign.Persistence.Systems.Persistence;
 
@@ -23,6 +27,15 @@ namespace Sovereign.Persistence.Systems.Persistence;
 /// </summary>
 public class PersistenceInternalController
 {
+    private readonly ILogger<PersistenceInternalController> logger;
+    private readonly NameComponentCollection names;
+
+    public PersistenceInternalController(NameComponentCollection names, ILogger<PersistenceInternalController> logger)
+    {
+        this.names = names;
+        this.logger = logger;
+    }
+
     /// <summary>
     ///     Announces that synchronization is complete.
     /// </summary>
@@ -30,6 +43,22 @@ public class PersistenceInternalController
     public void CompleteSync(IEventSender eventSender)
     {
         var ev = new Event(EventId.Server_Persistence_SynchronizeComplete);
+        eventSender.SendEvent(ev);
+    }
+
+    /// <summary>
+    ///     Announces that the data load is complete for a player entering the world.
+    /// </summary>
+    /// <param name="eventSender">Event sender.</param>
+    /// <param name="playerEntityId">Player entity ID.</param>
+    public void PlayerEnteredWorld(IEventSender eventSender, ulong playerEntityId)
+    {
+        var details = new EntityEventDetails
+        {
+            EntityId = playerEntityId
+        };
+        var ev = new Event(EventId.Server_Persistence_PlayerEnteredWorld, details);
+        ev.SyncToTick = true;
         eventSender.SendEvent(ev);
     }
 }