+++ src/Server/SovereignServer/Data/Scripts/Sovereign/Motd.lua
@@ -1,11 +1,6 @@
-g_ticks = 0
-
-function on_tick()
-    if g_ticks % 50 == 0 then
-        util.log_info("tick " .. g_ticks)
-    end
-    g_ticks = g_ticks + 1
+function on_player_entered(playerEntityId)
+    util.log_info("Player entered world: " .. playerEntityId)
 end
 
 util.log_info("This is a test of the scripting engine.")
-scripting.add_event_callback(events.Core_Tick, on_tick)
+scripting.add_event_callback(events.Server_Persistence_PlayerEnteredWorld, on_player_entered)